clean chat files front
