+++ frontend/src/components/chat/ChatAvailableRooms.vue
