+++ frontend/src/components/chat/ChatSelectedRoomParams.vue
