+++ frontend/src/router/index.ts
@@ -19,11 +19,11 @@ const router = createRouter({
     },
     {
       path: "/chat",
-      name: "ChatNew",
+      name: "Chat",
       // route level code-splitting
       // this generates a separate chunk (About.[hash].js) for this route
       // which is lazy-loaded when the route is visited.
-      component: () => import("../views/ChatNew.vue"),
+      component: () => import("../views/Chat.vue"),
     },
     {
       path: "/friends",
